Move JAX-RS classes from Model to Services
