+++ clients/deltalake/core/src/test/java/org/projectnessie/deltalake/ITDeltaLogBranches.java
@@ -127,7 +127,7 @@ class ITDeltaLogBranches extends AbstractSparkTest {
 
     String tableName = tempPath.getAbsolutePath() + "/_delta_log";
     Contents contents = client.getContentsApi()
-                              .getContents(new ContentsKey(Arrays.asList(tableName.split("/"))), "main");
+                              .getContents(ContentsKey.of(tableName.split("/")), "main");
     Optional<DeltaLakeTable> table = contents.unwrap(DeltaLakeTable.class);
     Assertions.assertTrue(table.isPresent());
     Assertions.assertEquals(1, table.get().getCheckpointLocationHistory().size());
